#381 - Refactor `/register-site` operation code
