+++ oxd-server/src/test/resources/oxd-server-dev.yml
@@ -78,9 +78,9 @@ defaultSiteConfig:
   client_jwks_uri: ''
   token_endpoint_auth_method: ''
   token_endpoint_auth_signing_alg: ''
-  client_request_uris: []
+  request_uris: []
   front_channel_logout_uris: []
-  client_sector_identifier_uri: ''
+  sector_identifier_uri: ''
   claims_redirect_uri: []
   client_id: ''
   client_secret: ''
